feat your tracks uploadheader
